Crate flix_ns

OmniFlix NameService

Modules

config

Contract Config Info

contract

Contract Entry Points

error

Contract Errors

helpers

Helpers

msg

Messages

state

States

Structs

Config

Configuration State

InstantiateMsg

Example Json Payload for InstantiateMsg :

Enums

ContractError

Contract Errors

ExecuteMsg

ExecuteMsg is the main JSON serializable enum. Based on its variant, execute will handle the execution. Its variants are the main entry to various end points.

QueryMsg

Query enum variants are for distinguishing various kinds of query requests

QueryResponse

Various Query Responses

Constants

CONFIG

Config state instance where the configuration info is stored.

MAX_NAME_LENGTH

Maximum length of the name in u64 characters

MIN_NAME_LENGTH

Minimum length of name in u64 characters

SUFFIX

The suffix for the name

Traits

Cw721Query

CW721 Contract Intigratation Variables

Functions

execute

The main Entry Point for the contract execution. It is called by the contract when an ExcuteMsg message is passed as the payload.

instantiate

query

Type Definitions

Cw721Contract

cw721 contract initialised using generics

Cw721ExecuteMsg

cw721 ExecuteMsg with generic over Extension

Cw721MintMsg

cw721 base provides MetaData extension for MintMsg as well

Extension

Cw721Contract allows metadata to be stored as an Extension. As there is no additional metadata that needs to be stored, it is Generic Empty